+++ mod.ts
@@ -4,3 +4,5 @@ export { TgaLoader };
 export default TgaLoader;
 export * from "./src/errors.js";
 export * from "./src/tga.js";
+export * from "./src/types.js";
+export * from "./src/writer.js";